requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
audioread==3.0.1
beautifulsoup4==4.12.3
bert-extractive-summarizer==0.10.1
blinker==1.8.2
blis==0.7.11
boto3==1.34.148
botocore==1.34.148
cachetools==5.4.0
catalogue==2.0.10
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
confection==0.1.5
coverage==7.6.1
coverage-badge==1.1.2
ctransformers==0.2.27
cymem==2.0.8
dataclasses-json==0.6.7
decorator==5.1.1
distro==1.9.0
dnspython==2.6.1
docstring_parser==0.16
email_validator==2.2.0
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.7.1/en_core_web_lg-3.7.1-py3-none-any.whl#sha256=ab70aeb6172cde82508f7739f35ebc9918a3d07debeed637403c8f794ba3d3dc
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
environs==9.5.0
fastapi==0.111.1
fastapi-cli==0.0.4
filelock==3.15.4
fire==0.6.0
frozenlist==1.4.1
fsspec==2024.6.1
gensim==4.3.3
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
grpcio==1.63.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
instructor==1.3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
jiter==0.4.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
langchain==0.2.7
langchain-community==0.2.7
langchain-core==0.2.16
langchain-text-splitters==0.2.2
langcodes==3.4.0
langsmith==0.1.85
language_data==1.2.0
lazy_loader==0.4
librosa==0.10.2.post1
llmware==0.3.4
llvmlite==0.43.0
lxml==5.2.2
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
mistral_common==1.3.3
mkl==2021.4.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
mysql-connector-python==9.0.0
networkx==3.3
nltk==3.8.1
numba==0.60.0
numpy==1.26.4
openai==1.37.1
orjson==3.10.6
packaging==24.1
pandas==2.2.2
pgvector==0.2.4
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
preshed==3.0.9
protobuf==5.27.2
psycopg==3.1.17
psycopg-binary==3.1.17
py-cpuinfo==9.0.0
pyarrow==17.0.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
pymilvus==2.4.4
pymongo==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.19.1
s3transfer==0.10.2
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.13.1
sentence-transformers==3.0.1
sentencepiece==0.2.0
setuptools==70.3.0
shellingham==1.5.4
simple_parsing==0.1.5
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.31
srsly==2.4.8
starlette==0.37.2
streamlit==1.36.0
sympy==1.13.0
tbb==2021.13.0
tenacity==8.5.0
termcolor==2.4.0
thinc==8.2.5
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.3.1
torch-directml==0.2.2.dev240614
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.4
transformers==4.42.4
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.3
wasabi==1.1.3
watchdog==4.0.1
watchfiles==0.22.0
weasel==0.4.1
websockets==12.0
wrapt==1.16.0
xformers==0.0.27
yarl==1.9.4
